+++ tests/test_all.py
@@ -1,53 +1,17 @@
 from unittest import TestCase
 
-expected_api = [
-    "AttrNode",
-    "AttrsNode",
-    "CliEngine",
-    "Config",
-    "ConfigList",
-    "DataArrayNode",
-    "DatasetNode",
-    "EditInfo",
-    "Formatter",
-    "FormatterContext",
-    "FormatterMeta",
-    "FormatterOp",
-    "FormatterRegistry",
-    "Linter",
-    "Message",
-    "Node",
-    "Plugin",
-    "PluginMeta",
-    "Processor",
-    "ProcessorMeta",
-    "ProcessorOp",
-    "Result",
-    "Rule",
-    "RuleConfig",
-    "RuleContext",
-    "RuleMeta",
-    "RuleOp",
-    "RuleTest",
-    "RuleTester",
-    "Suggestion",
-    "get_rules_meta_for_results",
-    "new_linter",
-    "version",
-]
-
 
 class AllTest(TestCase):
     def test_api_is_complete(self):
         import xrlint.all as xrl
 
+        # noinspection PyProtectedMember
+        from xrlint.all import __all__
+
         # noinspection PyUnresolvedReferences
-        keys = sorted(
+        keys = set(
             k
             for k, v in xrl.__dict__.items()
             if isinstance(k, str) and not k.startswith("_")
         )
-        self.assertEqual(
-            expected_api,
-            keys,
-        )
+        self.assertEqual(set(__all__), keys)